:basketball:

+++ leetcode-new/UniqueBinarySearchTrees.java
@@ -16,7 +16,7 @@ public class UniqueBinarySearchTrees {
     public int numTrees(int n) {
         int result = 1;
         for (int i = 0; i < n; ++i) {
-            result = result * 2 * (2 * i + 1) / (i + 2);
+            result = (int) ((long) result * 2 * (2 * i + 1) / (i + 2));
         }
         return result;
     }
